Fix Lesson 1

+++ lesson-01/perfect_weight.rb
@@ -1,7 +1,7 @@
 puts "Как Вас зовут?"
 name = gets.chomp
 puts "Какой у Вас рост?"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 perfect_weight = height - 110
 message = (perfect_weight < 0) ? "Ваш вес уже оптимальный" : "Ваш идеальный вес #{perfect_weight}"
